backend/models/document_processor.py
import os
from typing import List, Dict
import magic
from docx import Document
from PyPDF2 import PdfReader
from pydub import AudioSegment
import json
from PIL import Image
import torch
from transformers import AutoTokenizer, AutoModel

class DocumentProcessor:
    def __init__(self, upload_dir: str = "uploads"):
        self.upload_dir = upload_dir
        self.tokenizer = AutoTokenizer.from_pretrained("sentence-transformers/all-MiniLM-L6-v2")
        self.model = AutoModel.from_pretrained("sentence-transformers/all-MiniLM-L6-v2")
        os.makedirs(upload_dir, exist_ok=True)

    async def process_file(self, file_path: str) -> Dict:
        mime = magic.Magic(mime=True)
        file_type = mime.from_file(file_path)
        
        if "pdf" in file_type:
            return await self.process_pdf(file_path)
        elif "word" in file_type:
            return await self.process_docx(file_path)
        elif "audio" in file_type:
            return await self.process_audio(file_path)
        elif "image" in file_type:
            return await self.process_image(file_path)
        elif "video" in file_type:
            return await self.process_video(file_path)
        else:
            raise ValueError(f"Unsupported file type: {file_type}")

    async def process_pdf(self, file_path: str) -> Dict:
        reader = PdfReader(file_path)
        text = ""
        for page in reader.pages:
            text += page.extract_text()
        return {"text": text, "type": "pdf"}

    async def process_docx(self, file_path: str) -> Dict:
        doc = Document(file_path)
        text = "\n".join([paragraph.text for paragraph in doc.paragraphs])
        return {"text": text, "type": "docx"}

    async def process_audio(self, file_path: str) -> Dict:
        # Implement audio processing
        return {"type": "audio", "duration": ""}

    async def process_image(self, file_path: str) -> Dict:
        image = Image.open(file_path)
        return {"type": "image", "size": image.size}

    async def process_video(self, file_path: str) -> Dict:
        # Implement video processing
        return {"type": "video", "duration": ""}