backend/requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
python-multipart==0.0.9
websockets==12.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0
qdrant-client==1.7.0
python-magic==0.4.27
python-docx==1.1.0
PyPDF2==3.0.1
pydub==0.25.1
pillow==10.2.0
transformers==4.37.2
torch==2.2.0
numpy==1.24.3
python-multipart==0.0.9
google-auth==2.27.0
google-auth-oauthlib==1.2.0